// src/types/lastfm.ts
export type LastFMUser = string;

export enum LastFMPeriod {
  overall = "overall",
  sevenDays = "7day",
  oneMonth = "1month",
  threeMonths = "3month",
  sixMonths = "6month",
  twelveMonths = "12month",
}

export enum LastFMGridSize {
  three = "3",
  four = "4",
  five = "5",
  ten = "10",
}

export enum LastFMLimit {
  nine = "9",
  sixteen = "16",
  twentyFive = "25",
  oneHundred = "100",
}